requirements.txt
Flask>=3.1,<4.0
SQLAlchemy>=2.0,<3.0
Flask-SQLAlchemy>=3.1,<4.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
flask-cors>=6.0,<7.0
Flask-JWT-Extended>=4.7,<5.0
flask-restx>=1.3,<2.0
pandas>=2.2,<3.0
PyMuPDF>=1.26,<2.0
pillow>=11.0,<12.0
gunicorn>=23.0,<24.0
alembic>=1.14,<2.0
werkzeug>=3.0,<4.0
